requirements.txt
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
donfig==0.8.1.post1
evdev==1.9.3
fonttools==4.61.1
google-crc32c==1.8.0
kiwisolver==1.4.9
matplotlib==3.10.8
natnet==0.2.0
numcodecs==0.16.5
numpy==2.4.2
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scipy==1.17.0
six==1.17.0
typing_extensions==4.15.0
ur_rtde==1.6.2
zarr==3.1.5
